backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mss
pynput
python-dotenv
openai
Pillow
pygetwindow
pyobjc-framework-Quartz
